+++ pyvisim/retrieval/index/__init__.py
@@ -1,0 +1,6 @@
+"""Image indexes for accelerated similarity search."""
+
+from ._base_index import ImageIndex
+from .image_index import ImageIndexIVFFlat, ImageIndexIVFPQ
+
+__all__ = ["ImageIndex", "ImageIndexIVFFlat", "ImageIndexIVFPQ"]